+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## 0.6.0 (future)
+## 0.6.0
 
-* Schema changes: remove column `url` from `Tupperhook` model, replaced by `webhook_id`.  In previous codebase (not published), it used to store strings like `<Webhook id=123456789012345678>`.  And I found out now.
+* Added `/bibbia` command.
+* Schema changes: remove column `url` from `Tupperhook` model, replaced by `webhook_id`.  (In previous codebase, not published, it used to store strings like `<Webhook id=123456789012345678>`.  And I found out now.)
+* Now commands that fail have to return false in function body.  This way, commands that have failed do not give experience to player.
 
 ## 0.5.1